Add overs bowled to player bowling figures #116

+++ Stoolball/Statistics/BowlingStatistics.cs
@@ -4,7 +4,7 @@
     {
         public int TotalInnings { get; set; }
         public int TotalInningsWithRunsConceded { get; set; }
-        public int TotalOvers { get; set; }
+        public decimal TotalOvers { get; set; }
         public int TotalMaidens { get; set; }
         public int TotalRunsConceded { get; set; }
         public int TotalWickets { get; set; }
